requirement.txt
fastapi
uvicorn[standard]

sqlalchemy
asyncpg
alembic
psycopg2

python-jose[cryptography]
passlib[bcrypt]

pydantic
pydantic-settings

python-dotenv
httpx

loguru

PyJWT

python-docx
python-multipart